% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mirror_repo.R
\name{add_mirror_repo}
\alias{add_mirror_repo}
\title{Sets up a secondary git remote to push to}
\usage{
add_mirror_repo(mirror_repo, current_repo = ".", overwrite = FALSE)
}
\arguments{
\item{mirror_repo}{The repository that will serve as the mirror.  Maybe a 
GitHub repo URL or a local file path.  If a local file 
path, it must not already exist.  Do not use the .git 
extension}

\item{current_repo}{The repo you wish to mirror.  Default is to use the current
folder.  It must be a git repository.}
}
\description{
This function looks at a current git repository, and sets up two push 
remotes, as origin, to push to.  First push remote is the original origin and
second is a new, mirrored git repo.  The new repo serves as a mirror of the
original and is set up as a bare repo.  Because it is a bare repo you wont be 
able to browse the files and folders in the mirror location.  A clone of this 
repo will however restore the files and folders at the time of the last push.
}
\examples{
library(epar)
mirror <- one_drive_mirror("projects")
add_mirror_repo(mirror)                     
}
