% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choro.R
\name{choro}
\alias{choro}
\title{Choropleth mapping function}
\usage{
choro(epa_data, map_labels = FALSE)
}
\arguments{
\item{epa_data}{ETA TRI dataset for a state and year}

\item{map_labels}{Add county labels to the map (TRUE, FALSE)}
}
\value{
a ggplot object
}
\description{
A function for mapping chemical releases
}
\details{
This function takes in EPA data. It then creates a choropleth map using a shapefile joined with the data.
}
\examples{
\dontrun{
choro(ma_2024, TRUE)
choro(tx_2024, FALSE)
}
}
