% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_states.R
\name{compare_states}
\alias{compare_states}
\title{Comparing summary statistics between two states}
\usage{
compare_states(data, data2)
}
\arguments{
\item{data}{ETA TRI dataset for a state and year}

\item{data2}{ETA TRI dataset for another state and year}
}
\value{
text summarizing differences
}
\description{
A function for comparing statistics of two states
}
\details{
The user can specify two data sets for different states and the function will return three comparisons between the two: the number of releases, the percentage of hazardous chemicals, and the percentage of carcinogenic chemicals.
}
\examples{
\dontrun{
compare_states(ma_2024, tx_2024)

}
}
