% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_plot.R
\name{county_plot}
\alias{county_plot}
\title{County plotting function}
\usage{
county_plot(data, my_county)
}
\arguments{
\item{data}{ETA TRI dataset for a state and year}

\item{my_county}{County of interest}
}
\value{
a ggplot object
}
\description{
A function for plotting the number of facilities by industry
}
\details{
This function takes in EPA data and a county of interest. It filters the data to the specified county and groups by the industry sector, creating a plot of the number of chemical releases by facility in that county.
}
\examples{
\dontrun{
county_plot(ma_2024, "ESSEX")

}
}
