% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facil.R
\name{facil}
\alias{facil}
\title{Mapping function}
\usage{
facil(data, props = c(1, 2, 3))
}
\arguments{
\item{data}{ETA TRI dataset for a state and year}

\item{props}{Categorical variable to group by ("1", "2", "3"), depending on what the user wants the map to look like.}
}
\value{
a ggplot object
}
\description{
A mapping function that visualizes what \% of chemical releases are carcinogens by factory
}
\details{
This function takes in EPA data and graphs a simple map of the state and plots each facility. Then, according to what the user specifies, the points are either colored(1), sized(2), or both(3) according to the percentage of chemical releases that are carcinogens.
}
\examples{
\dontrun{
facil(ma_2024, 3)

}
}
