% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indust.R
\name{indust}
\alias{indust}
\title{Industry information}
\usage{
indust(data, grouping_var = "carcinogen", return_all = TRUE)
}
\arguments{
\item{data}{ETA TRI dataset for a state and year}

\item{grouping_var}{Binary variable to group by ("carcinogen", "pfas", "pbt", "federal_facility", "clean_air_act_chemical", "elemental_element_included", "metal")}

\item{return_all}{Determines whether the function returns a full or truncated table ("TRUE" for full, or a number for the top X rows)}
}
\value{
tibble
}
\description{
A function for exploring properties of different industry sectors in a given state.
}
\details{
The user can provide EPA data and a binary grouping variable of interest. The function will return the variable proportion for each industry sector, while also specifying the number of releases.
}
\examples{
\dontrun{
indust(ma_2024, "pbt")
}
}
