% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mirror_repo.R
\name{is_git}
\alias{is_git}
\title{Is this a git repo}
\usage{
is_git(repo = ".")
}
\arguments{
\item{repo}{Path to the folder to test if it is or is not a git repository.}
}
\description{
Function that returns TRUE or FALSE if the specified folder is a git repo or not.
}
