% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ma_2024}
\alias{ma_2024}
\title{ma_2024: EPA TRI data from Massachusetts}
\format{
A data frame with 820 rows and 122 variables:
\describe{
\item{test}{test}
}
}
\usage{
ma_2024
}
\description{
Contains chemical release data from Massachusetts in 2024. Can be used as an example data set for any package functions.
}
\keyword{datasets}
