% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapchems.R
\name{mapchems}
\alias{mapchems}
\title{Simple mapping function}
\usage{
mapchems(data, grouping_var = "classification")
}
\arguments{
\item{data}{ETA TRI dataset for a state and year}

\item{grouping_var}{Categorical variable to group by ("classification", "carcinogen", "federal_facility", "pbt", "pfas")}
}
\value{
a ggplot object
}
\description{
A function for mapping chemical releases
}
\details{
This function takes in EPA data and a grouping variable. It then creates a simple geographic scatter plot using lat/lon coordinates, differentiating by the grouping variable.
}
\examples{
\dontrun{
mapchems(ma_2024, "classification")

}
}
