% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.R
\name{percentage}
\alias{percentage}
\title{Getting the percentage of how many values of a column are equal to a specific value}
\usage{
percentage(data, column, value)
}
\arguments{
\item{data}{ETA TRI dataset for a state and year}

\item{column}{Name of a column in data}

\item{value}{Specific value to count}
}
\value{
percentage
}
\description{
A function for getting the percentage of instances of a specific value
}
\details{
The user can give a dataset, a column, and a specific value and the function will return the percentage of values in the column that is equal to a specific value.
}
\examples{
\dontrun{
percentage(ma_2024, "metal", "YES")

}
}
