% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_and_badge.R
\name{test_and_badge}
\alias{test_and_badge}
\title{Run package tests and create badge}
\usage{
test_and_badge(package = ".")
}
\arguments{
\item{package}{path to package}
}
\description{
This function uses the \code{covr} package to run tests for a package and 
returns a markdown badge code reporting the code coverage that can be used on
a README (or any markdown).
}
