% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_counties}
\alias{us_counties}
\title{us_counties: shapefile for the United States, by state counties}
\format{
A data frame with 3235 rows and 6 variables:
\describe{
\item{STATEFP}{Unique state code}
\item{COUNTYFP}{Unique county code}
\item{county}{Name of county, lowercase}
\item{STUSPS}{State abbreviation}
\item{STATE_NAME}{State name}
\item{geometry}{Spatial geometry containing latitude and longitude coordinates}
}
}
\usage{
us_counties
}
\description{
Contains shapefile information (spatial geometry) needed for mapping functions within the package, such as mapchems() and choro().
}
\keyword{datasets}
