% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemstats.R
\name{chemstats}
\alias{chemstats}
\title{Statistic summary by area}
\usage{
chemstats(data, area = c("county", "city", "zip"))
}
\arguments{
\item{data}{ETA TRI dataset for a state and year}

\item{area}{User choice of county, city or zip}
}
\value{
summary statistics
}
\description{
A function for summarizing statistics of a given area
}
\details{
The user can specify county, city, or zip code, and from there this function will return the number of releases, the percentage of carcinogens, the percentage of hazardous chemicals, and the top industry by counties/cities/zips.
}
\examples{
\dontrun{
chemstats(epa, "county")

}
}
