% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_epa.R
\name{clean_epa}
\alias{clean_epa}
\title{Cleaning data function}
\usage{
clean_epa(data)
}
\arguments{
\item{data}{ETA TRI dataset for a state and year}
}
\value{
a data frame
}
\description{
A function that cleans any data set from the EPA TRI program website.
}
\details{
This function cleans up the inputted data by updating column names to more usable formats.
}
\examples{
\dontrun{
clean_epa(epa_data)

}
}
