% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_change.R
\name{unit_change}
\alias{unit_change}
\title{Changing units}
\usage{
unit_change(epa_data, new_unit)
}
\arguments{
\item{epa_data}{ETA TRI dataset for a state and year}

\item{new_unit}{User choice of lbs, kg, or g}
}
\value{
a dataset
}
\description{
A function for changing the units of the continuous total release information.
}
\details{
This is a function that can be used to change and synthesize the units of the total_release variable of the epa data. Original units are in pounds, other than dioxins which are in grams. This function can convert the dioxins to pounds, and then convert everything to kilograms or grams if desired.
}
\examples{
\dontrun{
unit_change(epa_data, "lbs")

}
}
